Moved TimeSeries to it's own class - and removed it from base
